requirements.txt
Flask>=2.0
flask-cors
python-dotenv
pypdf
fpdf
streamlit
langchain
langchain-core
openai